@@ -1,425 +1,488 @@
-<?xml version="1.0" encoding="UTF-8" ?>
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  Refactored POM for taxonomy-api (Java 21, Jakarta EE 10, Jersey 3.x, OpenAPI 3.x)
+  Use this as a drop-in replacement for your current POM.
+-->
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<groupId>com.strandls.taxonomy</groupId>
-	<artifactId>taxonomy-api</artifactId>
-	<version>1.0.0</version>
-	<packaging>war</packaging>
-	<name>biodiv-taxonomy</name>
-	<url>https://github.com/strandls/biodiv-taxonomy</url>
+    <modelVersion>4.0.0</modelVersion>
+    <groupId>com.strandls.taxonomy</groupId>
+    <artifactId>taxonomy-api</artifactId>
+    <version>1.0.0</version>
+    <packaging>war</packaging>
+    <name>biodiv-taxonomy</name>
+    <url>https://github.com/strandls/biodiv-taxonomy</url>
 
-	<properties>
+    <properties>
+		<!-- Java/Compiler -->
+        <java.version>21</java.version>
+        <maven.compiler.release>21</maven.compiler.release>
 
-		<!-- Plugins property -->
-		<maven-compiler-plugin.version>3.8.0</maven-compiler-plugin.version>
-		<maven-dependency-plugin.version>3.1.1</maven-dependency-plugin.version>
-		<swagger-ui.version>3.23.5</swagger-ui.version>
-		<maven-war-plugin.version>3.2.2</maven-war-plugin.version>
-		<replacer.version>1.5.3</replacer.version>
-		<swagger-maven-plugin.version>3.1.1</swagger-maven-plugin.version>
-		<exec-maven-plugin.version>1.3.2</exec-maven-plugin.version>
-		<openapi-generator-cli.version>4.0.3</openapi-generator-cli.version>
+		<!-- Plugin versions -->
+        <maven-compiler-plugin.version>3.8.0</maven-compiler-plugin.version>
+        <maven-dependency-plugin.version>3.1.1</maven-dependency-plugin.version>
+        <maven-war-plugin.version>3.2.2</maven-war-plugin.version>
+        <swagger-ui.version>3.23.5</swagger-ui.version>
+        <swagger.version>2.2.21</swagger.version>
+        <openapi-generator-cli.version>7.2.0</openapi-generator-cli.version>
+        <replacer.version>1.5.3</replacer.version>
+        <exec-maven-plugin.version>3.1.0</exec-maven-plugin.version>
+        <hibernate-core.version>6.2.7.Final</hibernate-core.version>
 
-		<!-- Project property -->
-		<war-name>taxonomy-api</war-name>
-		<srcDir>src/main/webapp</srcDir>
-		<base-path>/taxonomy-api/api</base-path>
-		<location>com.strandls.taxonomy</location>
+        <war-name>taxonomy-api</war-name>
+        <srcDir>src/main/webapp</srcDir>
+        <base-path>/taxonomy-api/api</base-path>
+        <location>com.strandls.taxonomy</location>
 
-		<!-- Sonar Cloud -->
-		<sonar.projectKey>strandls_biodiv-taxonomy</sonar.projectKey>
-		<sonar.organization>strandls</sonar.organization>
-		<sonar.host.url>https://sonarcloud.io</sonar.host.url>
-	</properties>
+		<!-- SonarCloud -->
+        <sonar.projectKey>strandls_biodiv-taxonomy</sonar.projectKey>
+        <sonar.organization>strandls</sonar.organization>
+        <sonar.host.url>https://sonarcloud.io</sonar.host.url>
+    </properties>
 
-	<distributionManagement>
-		<repository>
-			<id>central</id>
-			<name>jfrog-releases</name>
-			<url>https://devartifactory.indiabiodiversity.org/libs-release</url>
-		</repository>
-		<snapshotRepository>
-			<id>snapshots</id>
-			<name>jfrog-snapshots</name>
-			<url>https://devartifactory.indiabiodiversity.org/libs-snapshot</url>
-		</snapshotRepository>
-	</distributionManagement>
+    <distributionManagement>
+        <repository>
+            <id>central</id>
+            <name>jfrog-releases</name>
+            <url>https://devartifactory.indiabiodiversity.org/libs-release</url>
+        </repository>
+        <snapshotRepository>
+            <id>snapshots</id>
+            <name>jfrog-snapshots</name>
+            <url>https://devartifactory.indiabiodiversity.org/libs-snapshot</url>
+        </snapshotRepository>
+    </distributionManagement>
 
-	<repositories>
-		<repository>
-			<snapshots>
-				<enabled>false</enabled>
-			</snapshots>
-			<id>central</id>
-			<name>libs-release</name>
-			<url>https://devartifactory.indiabiodiversity.org/libs-release</url>
-		</repository>
-		<repository>
-			<snapshots />
-			<id>snapshots</id>
-			<name>libs-snapshot</name>
-			<url>https://devartifactory.indiabiodiversity.org/libs-snapshot</url>
-		</repository>
-	</repositories>
+    <repositories>
+        <repository>
+            <id>central</id>
+            <name>libs-release</name>
+            <url>https://devartifactory.indiabiodiversity.org/libs-release</url>
+            <snapshots>
+                <enabled>false</enabled>
+            </snapshots>
+        </repository>
+        <repository>
+            <id>snapshots</id>
+            <name>libs-snapshot</name>
+            <url>https://devartifactory.indiabiodiversity.org/libs-snapshot</url>
+            <snapshots>
+                <enabled>true</enabled>
+            </snapshots>
+        </repository>
+    </repositories>
 
-	<dependencies>
+    <dependencies>
+		<!-- Logging -->
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+            <version>2.0.7</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.logging.log4j</groupId>
+            <artifactId>log4j-core</artifactId>
+            <version>2.20.0</version>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-log4j12</artifactId>
+            <version>2.0.7</version>
+        </dependency>
 
-		<!-- https://mvnrepository.com/artifact/org.slf4j/slf4j-api -->
-		<dependency>
-			<groupId>org.slf4j</groupId>
-			<artifactId>slf4j-api</artifactId>
-			<version>1.7.13</version>
-		</dependency>
+		<!-- Jakarta EE 10 / Jersey 3.x -->
+        <dependency>
+            <groupId>jakarta.ws.rs</groupId>
+            <artifactId>jakarta.ws.rs-api</artifactId>
+            <version>3.1.0</version>
+        </dependency>
+        <dependency>
+            <groupId>org.glassfish.jersey.core</groupId>
+            <artifactId>jersey-server</artifactId>
+            <version>3.1.2</version>
+        </dependency>
+        <dependency>
+            <groupId>org.glassfish.jersey.containers</groupId>
+            <artifactId>jersey-container-servlet-core</artifactId>
+            <version>3.1.2</version>
+        </dependency>
+        <dependency>
+            <groupId>org.glassfish.jersey.inject</groupId>
+            <artifactId>jersey-hk2</artifactId>
+            <version>3.1.2</version>
+        </dependency>
+        <dependency>
+            <groupId>org.glassfish.hk2</groupId>
+            <artifactId>guice-bridge</artifactId>
+            <version>3.0.0</version>
+        </dependency>
+        <dependency>
+            <groupId>com.google.inject</groupId>
+            <artifactId>guice</artifactId>
+            <version>7.0.0</version>
+        </dependency>
+        <dependency>
+            <groupId>com.google.inject.extensions</groupId>
+            <artifactId>guice-servlet</artifactId>
+            <version>7.0.0</version>
+        </dependency>
+        <dependency>
+            <groupId>jakarta.platform</groupId>
+            <artifactId>jakarta.jakartaee-web-api</artifactId>
+            <version>10.0.0</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.glassfish.jersey.media</groupId>
+            <artifactId>jersey-media-json-jackson</artifactId>
+            <version>3.1.2</version>
+        </dependency>
+        <dependency>
+            <groupId>org.glassfish.jersey.media</groupId>
+            <artifactId>jersey-media-multipart</artifactId>
+            <version>3.1.2</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-core</artifactId>
+            <version>2.15.2</version>
+        </dependency>
 
-		<!-- https://mvnrepository.com/artifact/org.apache.logging.log4j/log4j-core -->
-		<dependency>
-			<groupId>org.apache.logging.log4j</groupId>
-			<artifactId>log4j-core</artifactId>
-			<version>2.17.1</version>
-		</dependency>
+		<!-- JAXB (Java 21+) -->
+        <dependency>
+            <groupId>org.glassfish.jaxb</groupId>
+            <artifactId>jaxb-runtime</artifactId>
+            <version>4.0.4</version>
+        </dependency>
 
-		<dependency>
-			<groupId>org.slf4j</groupId>
-			<artifactId>slf4j-log4j12</artifactId>
-			<version>1.7.13</version>
-		</dependency>
+		<!-- OpenAPI 3.x / Swagger 3 -->
+        <dependency>
+            <groupId>io.swagger.core.v3</groupId>
+            <artifactId>swagger-jaxrs2-jakarta</artifactId>
+            <version>${swagger.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.swagger.core.v3</groupId>
+            <artifactId>swagger-integration</artifactId>
+            <version>${swagger.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.swagger.core.v3</groupId>
+            <artifactId>swagger-annotations-jakarta</artifactId>
+            <version>${swagger.version}</version>
+        </dependency>
 
+		<!-- Database/ORM -->
+        <dependency>
+            <groupId>org.hibernate.orm</groupId>
+            <artifactId>hibernate-core</artifactId>
+            <version>${hibernate-core.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.hibernate</groupId>
+            <artifactId>hibernate-hikaricp</artifactId>
+            <version>${hibernate-core.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.zaxxer</groupId>
+            <artifactId>HikariCP</artifactId>
+            <version>3.4.2</version>
+        </dependency>
+        <dependency>
+            <groupId>org.postgresql</groupId>
+            <artifactId>postgresql</artifactId>
+            <version>42.3.3</version>
+        </dependency>
+        <dependency>
+            <groupId>com.opencsv</groupId>
+            <artifactId>opencsv</artifactId>
+            <version>5.9</version>
+        </dependency>
+        <dependency>
+            <groupId>com.vladmihalcea</groupId>
+            <artifactId>hibernate-types-60</artifactId>
+            <version>2.21.1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.rabbitmq</groupId>
+            <artifactId>amqp-client</artifactId>
+            <version>5.18.0</version>
+        </dependency>
 
-	<!-- https://mvnrepository.com/artifact/javax.ws.rs/javax.ws.rs-api -->
-		<dependency>
-			<groupId>javax.ws.rs</groupId>
-			<artifactId>javax.ws.rs-api</artifactId>
-			<version>2.1.1</version>
-		</dependency>
+		<!-- StrandLS/IBP Clients -->
+        <dependency>
+            <groupId>com.strandls.biodiv</groupId>
+            <artifactId>authentication-utility</artifactId>
+            <version>0.0.21</version>
+        </dependency>
+        <dependency>
+            <groupId>com.strandls.biodiv</groupId>
+            <artifactId>user-client</artifactId>
+            <version>0.0.21</version>
+        </dependency>
+        <dependency>
+            <groupId>com.strandls.biodiv</groupId>
+            <artifactId>utility-client</artifactId>
+            <version>0.0.21</version>
+        </dependency>
+        <dependency>
+            <groupId>com.strandls.biodiv</groupId>
+            <artifactId>esmodule-client</artifactId>
+            <version>0.0.21</version>
+        </dependency>
+        <dependency>
+            <groupId>com.strandls.biodiv</groupId>
+            <artifactId>activity-client</artifactId>
+            <version>0.0.21</version>
+        </dependency>
+        <dependency>
+            <groupId>com.strandls.biodiv</groupId>
+            <artifactId>mail-utility</artifactId>
+            <version>0.0.21</version>
+        </dependency>
 
-		<!-- https://mvnrepository.com/artifact/org.glassfish.jersey.core/jersey-server -->
-		<dependency>
-			<groupId>org.glassfish.jersey.core</groupId>
-			<artifactId>jersey-server</artifactId>
-			<version>2.31</version>
-		</dependency>
+    </dependencies>
 
-		<!-- https://mvnrepository.com/artifact/org.glassfish.jersey.containers/jersey-container-servlet-core -->
-		<dependency>
-			<groupId>org.glassfish.jersey.containers</groupId>
-			<artifactId>jersey-container-servlet-core</artifactId>
-			<version>2.31</version>
-		</dependency>
+	<!-- =========================
+         Build Plugins
+         ========================= -->
+    <build>
+        <finalName>${war-name}</finalName>
+        <plugins>
+			<!-- Java Compiler Configuration -->
+            <plugin>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>${maven-compiler-plugin.version}</version>
+                <configuration>
+                    <release>${maven.compiler.release}</release>
+                </configuration>
+            </plugin>
 
-		<dependency>
-			<groupId>org.glassfish.jersey.inject</groupId>
-			<artifactId>jersey-hk2</artifactId>
-			<version>2.31</version>
-		</dependency>
+			<!-- Swagger Maven Plugin: generates swagger.json (OpenAPI documentation) -->
+            <plugin>
+                <groupId>io.swagger.core.v3</groupId>
+                <artifactId>swagger-maven-plugin-jakarta</artifactId>
+                <version>${swagger.version}</version>
+                <executions>
+                    <execution>
+                        <id>generate-openapi</id>
+                        <phase>compile</phase>
+                        <goals>
+                            <goal>resolve</goal>
+                        </goals>
+                        <configuration>
+                            <outputFileName>swagger</outputFileName>
+                            <outputFormat>JSON</outputFormat>
+                            <outputPath>${project.build.directory}</outputPath>
+                            <prettyPrint>true</prettyPrint>
+                            <resourcePackages>
+                                <package>${location}</package>
+                            </resourcePackages>
+                        </configuration>
+                    </execution>
+                </executions>
+				<!-- Required for servlet scanning -->
+                <dependencies>
+                    <dependency>
+                        <groupId>jakarta.servlet</groupId>
+                        <artifactId>jakarta.servlet-api</artifactId>
+                        <version>5.0.0</version>
+                    </dependency>
+                </dependencies>
+            </plugin>
 
-		<!-- https://mvnrepository.com/artifact/org.glassfish.hk2/guice-bridge -->
-		<dependency>
-			<groupId>org.glassfish.hk2</groupId>
-			<artifactId>guice-bridge</artifactId>
-			<version>2.6.1</version>
-		</dependency>
+			<!-- Exec Maven Plugin:
+                 1. Modifies generated swagger.json
+                 2. Generates SDK from swagger.json using OpenAPI Generator
+            -->
+            <plugin>
+                <groupId>org.codehaus.mojo</groupId>
+                <artifactId>exec-maven-plugin</artifactId>
+                <version>${exec-maven-plugin.version}</version>
+                <executions>
+                    <execution>
+                        <id>modify-openapi-json</id>
+                        <phase>compile</phase>
+                        <goals>
+                            <goal>java</goal>
+                        </goals>
+                        <configuration>
+                            <mainClass>com.strandls.taxonomy.ModifyOpenApi</mainClass>
+                            <arguments>
+                                <argument>${project.build.directory}/swagger.json</argument>
+                            </arguments>
+                        </configuration>
+                    </execution>
+                    <execution>
+                        <id>generate-sdk</id>
+                        <phase>package</phase>
+                        <goals>
+                            <goal>java</goal>
+                        </goals>
+                        <configuration>
+                            <mainClass>org.openapitools.codegen.OpenAPIGenerator</mainClass>
+                            <arguments>
+                                <argument>generate</argument>
+                                <argument>-i</argument>
+                                <argument>${project.build.directory}/swagger.json</argument>
+                                <argument>-g</argument>
+                                <argument>java</argument>
+                                <argument>-o</argument>
+                                <argument>target/sdk</argument>
+                                <argument>--skip-validate-spec</argument>
+                                <argument>-c</argument>
+                                <argument>swagger-config.json</argument>
+                            </arguments>
+                            <includeProjectDependencies>true</includeProjectDependencies>
+                            <includePluginDependencies>true</includePluginDependencies>
+                        </configuration>
+                    </execution>
+                </executions>
+                <dependencies>
+                    <dependency>
+                        <groupId>org.openapitools</groupId>
+                        <artifactId>openapi-generator-cli</artifactId>
+                        <version>${openapi-generator-cli.version}</version>
+                    </dependency>
+                </dependencies>
+            </plugin>
 
-		<!-- https://mvnrepository.com/artifact/com.google.inject/guice -->
-		<dependency>
-			<groupId>com.google.inject</groupId>
-			<artifactId>guice</artifactId>
-			<version>4.2.3</version>
-		</dependency>
+			<!-- Maven Resources Plugin:
+                 Copies modified swagger.json to swagger-ui folder in build directory
+            -->
+            <plugin>
+                <artifactId>maven-resources-plugin</artifactId>
+                <version>3.3.1</version>
+                <executions>
+                    <execution>
+                        <id>copy-openapi</id>
+                        <phase>prepare-package</phase>
+                        <goals>
+                            <goal>copy-resources</goal>
+                        </goals>
+                        <configuration>
+                            <outputDirectory>${project.build.directory}/swagger-ui</outputDirectory>
+                            <resources>
+                                <resource>
+                                    <directory>${project.build.directory}</directory>
+                                    <includes>
+                                        <include>swagger.json</include>
+                                    </includes>
+                                </resource>
+                            </resources>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
 
-		<!-- https://mvnrepository.com/artifact/com.google.inject.extensions/guice-servlet -->
-		<dependency>
-			<groupId>com.google.inject.extensions</groupId>
-			<artifactId>guice-servlet</artifactId>
-			<version>4.2.3</version>
-		</dependency>
+			<!-- Maven Dependency Plugin:
+                 Unpacks Swagger UI web resources from WebJars to swagger-ui folder
+            -->
+            <plugin>
+                <artifactId>maven-dependency-plugin</artifactId>
+                <version>${maven-dependency-plugin.version}</version>
+                <executions>
+                    <execution>
+                        <phase>prepare-package</phase>
+                        <goals>
+                            <goal>unpack</goal>
+                        </goals>
+                        <configuration>
+                            <artifactItems>
+                                <artifactItem>
+                                    <groupId>org.webjars</groupId>
+                                    <artifactId>swagger-ui</artifactId>
+                                    <version>${swagger-ui.version}</version>
+                                </artifactItem>
+                            </artifactItems>
+                            <outputDirectory>${project.build.directory}/swagger-ui</outputDirectory>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
 
-		<!-- https://mvnrepository.com/artifact/javax/javaee-web-api -->
-		<dependency>
-			<groupId>javax</groupId>
-			<artifactId>javaee-web-api</artifactId>
-			<version>8.0.1</version>
-			<scope>provided</scope>
-		</dependency>
+			<!-- Maven Replacer Plugin:
+                 Modifies Swagger UI's index.html to point to local swagger.json instead of demo petstore
+            -->
+            <plugin>
+                <groupId>com.google.code.maven-replacer-plugin</groupId>
+                <artifactId>replacer</artifactId>
+                <version>${replacer.version}</version>
+                <executions>
+                    <execution>
+                        <phase>prepare-package</phase>
+                        <goals>
+                            <goal>replace</goal>
+                        </goals>
+                    </execution>
+                </executions>
+                <configuration>
+                    <file>
+                        ${project.build.directory}/swagger-ui/META-INF/resources/webjars/swagger-ui/${swagger-ui.version}/index.html
+                    </file>
+                    <replacements>
+                        <replacement>
+                            <token>https://petstore.swagger.io/v2/swagger.json</token>
+                            <value>/taxonomy-api/swagger-ui/swagger.json</value>
+                        </replacement>
+                    </replacements>
+                </configuration>
+            </plugin>
 
-		<!-- https://mvnrepository.com/artifact/org.glassfish.jersey.media/jersey-media-json-jackson -->
-		<dependency>
-			<groupId>org.glassfish.jersey.media</groupId>
-			<artifactId>jersey-media-json-jackson</artifactId>
-			<version>2.31</version>
-		</dependency>
-		
-		<!-- https://mvnrepository.com/artifact/org.glassfish.jersey.media/jersey-media-multipart -->
-		<dependency>
-		    <groupId>org.glassfish.jersey.media</groupId>
-		    <artifactId>jersey-media-multipart</artifactId>
-		    <version>2.31</version>
-		</dependency>
-		
-		<!-- https://mvnrepository.com/artifact/com.fasterxml.jackson.core/jackson-core -->
-		<dependency>
-		    <groupId>com.fasterxml.jackson.core</groupId>
-		    <artifactId>jackson-core</artifactId>
-		    <version>2.11.0</version>
-		</dependency>
-				
-		<dependency>
-		    <groupId>com.sun.xml.bind</groupId>
-		    <artifactId>jaxb-core</artifactId>
-		    <version>2.3.0.1</version>
-		</dependency>
+			<!-- Maven WAR Plugin:
+                 Packages Swagger UI static files and swagger.json into the final WAR
+            -->
+            <plugin>
+                <artifactId>maven-war-plugin</artifactId>
+                <version>${maven-war-plugin.version}</version>
+                <configuration>
+                    <warName>${war-name}</warName>
+                    <warSourceDirectory>${srcDir}</warSourceDirectory>
+                    <webResources combine.children="append">
+						<!-- Bundle Swagger UI web resources -->
+                        <resource>
+                            <directory>
+                                ${project.build.directory}/swagger-ui/META-INF/resources/webjars/swagger-ui/${swagger-ui.version}</directory>
+                            <includes>
+                                <include>**/*.*</include>
+                            </includes>
+                            <targetPath>swagger-ui</targetPath>
+                        </resource>
+						<!-- Bundle the generated swagger.json -->
+                        <resource>
+                            <directory>${project.build.directory}/swagger-ui</directory>
+                            <includes>
+                                <include>swagger.json</include>
+                            </includes>
+                            <targetPath>swagger-ui</targetPath>
+                        </resource>
+                    </webResources>
+                </configuration>
+            </plugin>
 
-		<dependency>
-		    <groupId>com.sun.xml.bind</groupId>
-		    <artifactId>jaxb-impl</artifactId>
-		    <version>2.3.1</version>
-		</dependency>
-		
-		<!-- https://mvnrepository.com/artifact/io.swagger/swagger-jersey2-jaxrs -->
-		<dependency>
-			<groupId>io.swagger</groupId>
-			<artifactId>swagger-jersey2-jaxrs</artifactId>
-			<version>1.6.1</version>
-		</dependency>
-		
-		<!-- https://mvnrepository.com/artifact/io.swagger/swagger-annotations -->
-		<dependency>
-		    <groupId>io.swagger</groupId>
-		    <artifactId>swagger-annotations</artifactId>
-		    <version>1.6.1</version>
-		</dependency>
-		
-
-
-		<!-- https://mvnrepository.com/artifact/org.hibernate/hibernate-core -->
-		<dependency>
-			<groupId>org.hibernate</groupId>
-			<artifactId>hibernate-core</artifactId>
-			<version>5.4.24.Final</version>
-		</dependency>
-
-		<!-- https://mvnrepository.com/artifact/org.postgresql/postgresql -->
-		<dependency>
-			<groupId>org.postgresql</groupId>
-			<artifactId>postgresql</artifactId>
-			<version>42.3.3</version>
-		</dependency>
-		
-		 <!-- https://mvnrepository.com/artifact/org.hibernate/hibernate-hikaricp -->
-		<dependency>
-		    <groupId>org.hibernate</groupId>
-		    <artifactId>hibernate-hikaricp</artifactId>
-		    <version>5.2.12.Final</version>
-		</dependency>
-		
-		<!-- https://mvnrepository.com/artifact/com.zaxxer/HikariCP -->
-		<dependency>
-		    <groupId>com.zaxxer</groupId>
-		    <artifactId>HikariCP</artifactId>
-		    <version>3.4.2</version>
-		</dependency>
-		
-		<!-- https://mvnrepository.com/artifact/com.opencsv/opencsv -->
-		<dependency>
-		    <groupId>com.opencsv</groupId>
-		    <artifactId>opencsv</artifactId>
-		    <version>5.3</version>
-		</dependency>
-
-		<!-- https://mvnrepository.com/artifact/com.vladmihalcea/hibernate-types-52 -->
-		<dependency>
-			<groupId>com.vladmihalcea</groupId>
-			<artifactId>hibernate-types-52</artifactId>
-			<version>2.10.0</version>
-		</dependency>
-		
-		<dependency>
-		    <groupId>com.strandls.biodiv</groupId>
-		    <artifactId>authentication-utility</artifactId>
-		    <version>1.0.0</version>
-		</dependency>
-		
-		<dependency>
-			<groupId>com.strandls.biodiv</groupId>
-			<artifactId>user-client</artifactId>
-			<version>1.0.0</version>
-		</dependency>
-		
-		<dependency>
-			<groupId>com.strandls.biodiv</groupId>
-			<artifactId>utility-client</artifactId>
-			<version>1.0.0</version>
-		</dependency>
-		
-		<dependency>
-			<groupId>com.strandls.biodiv</groupId>
-			<artifactId>esmodule-client</artifactId>
-			<version>1.0.0</version>
-		</dependency>
-		
-		<dependency>
-		    <groupId>com.strandls.biodiv</groupId>
-		    <artifactId>activity-client</artifactId>
-		    <version>1.0.0</version>
-		</dependency>
-		
-		<!-- https://mvnrepository.com/artifact/com.rabbitmq/amqp-client -->
-		<dependency>
-			<groupId>com.rabbitmq</groupId>
-			<artifactId>amqp-client</artifactId>
-			<version>5.8.0</version>
-		</dependency>
-		
-		<dependency>
-		    <groupId>com.strandls.biodiv</groupId>
-		    <artifactId>mail-utility</artifactId>
-		    <version>1.0.0</version>
-		</dependency>
-		
-	</dependencies>
-
-	<build>
-		<finalName>taxonomy</finalName>
-		<plugins>
-			<plugin>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<version>${maven-compiler-plugin.version}</version>
-				<configuration>
-					<source>1.8</source>
-					<target>1.8</target>
-				</configuration>
-			</plugin>
-
-			<plugin>
-				<!-- Download Swagger UI webjar. -->
-				<artifactId>maven-dependency-plugin</artifactId>
-				<version>${maven-dependency-plugin.version}</version>
-				<executions>
-					<execution>
-						<phase>prepare-package</phase>
-						<goals>
-							<goal>unpack</goal>
-						</goals>
-						<configuration>
-							<artifactItems>
-								<artifactItem>
-									<groupId>org.webjars</groupId>
-									<artifactId>swagger-ui</artifactId>
-									<version>${swagger-ui.version}</version>
-								</artifactItem>
-							</artifactItems>
-							<outputDirectory> ${project.build.directory}/swagger-ui</outputDirectory>
-						</configuration>
-					</execution>
-				</executions>
-			</plugin>
-
-			<plugin>
-				<artifactId>maven-war-plugin</artifactId>
-				<version>${maven-war-plugin.version}</version>
-				<configuration>
-					<webResources combine.children="append">
-						<resource>
-							<directory>${project.build.directory}/swagger-ui/META-INF/resources/webjars/swagger-ui/${swagger-ui.version}</directory>
-							<includes>
-								<include>**/*.*</include>
-							</includes>
-							<targetPath>swagger-ui</targetPath>
-						</resource>
-					</webResources>
-					<warName>${war-name}</warName>
-					<warSourceDirectory>${srcDir}</warSourceDirectory>
-				</configuration>
-			</plugin>
-
-			<plugin>
-				<!-- Replace the OpenAPI specification example URL with the local one. -->
-				<groupId>com.google.code.maven-replacer-plugin</groupId>
-				<artifactId>replacer</artifactId>
-				<version>${replacer.version}</version>
-				<executions>
-					<execution>
-						<phase>prepare-package</phase>
-						<goals>
-							<goal>replace</goal>
-						</goals>
-					</execution>
-				</executions>
-				<configuration>
-					<file>${project.build.directory}/swagger-ui/META-INF/resources/webjars/swagger-ui/${swagger-ui.version}/index.html</file>
-					<replacements>
-						<replacement>
-							<token>https://petstore.swagger.io/v2/swagger.json</token>
-							<value>../api/swagger.json</value>
-						</replacement>
-					</replacements>
-				</configuration>
-			</plugin>
-
-			<plugin>
-				<groupId>com.github.kongchen</groupId>
-				<artifactId>swagger-maven-plugin</artifactId>
-				<version>${swagger-maven-plugin.version}</version>
-				<configuration>
-					<apiSources>
-						<apiSource>
-							<springmvc>false</springmvc>
-							<locations>
-								${location}
-							</locations>
-							<schemes>http</schemes>
-							<host>localhost:8080</host>
-							<basePath>${base-path}</basePath>
-							<info>
-								<title>Swagger file generation</title>
-								<version>v1</version>
-							</info>
-							<outputPath>${project.basedir}/target</outputPath>
-							<outputFormats>json</outputFormats>
-							<swaggerDirectory>${project.basedir}/target</swaggerDirectory>
-						</apiSource>
-					</apiSources>
-				</configuration>
-				<executions>
-					<execution>
-						<phase>test</phase>
-						<goals>
-							<goal>generate</goal>
-						</goals>
-					</execution>
-				</executions>
-			</plugin>
-
-			<plugin>
-				<groupId>org.codehaus.mojo</groupId>
-				<artifactId>exec-maven-plugin</artifactId>
-				<version>${exec-maven-plugin.version}</version>
-				<executions>
-					<execution>
-						<phase>package</phase>
-						<goals>
-							<goal>java</goal>
-						</goals>
-					</execution>
-				</executions>
-				<configuration>
-					<executable>java</executable>
-					<mainClass>org.openapitools.codegen.OpenAPIGenerator</mainClass>
-					<commandlineArgs>
-						generate -i ${project.basedir}/target/swagger.json -g java -o target/sdk --skip-validate-spec -c swagger-config.json
-					</commandlineArgs>
-					<includeProjectDependencies>true</includeProjectDependencies>
-					<includePluginDependencies>true</includePluginDependencies>
-				</configuration>
-				<dependencies>
-					<dependency>
-						<groupId>org.openapitools</groupId>
-						<artifactId>openapi-generator-cli</artifactId>
-						<version>${openapi-generator-cli.version}</version>
-					</dependency>
-				</dependencies>
-			</plugin>
-		</plugins>
-	</build>
-
+			<!-- Format java files by running 'mvn spotless:apply' -->
+            <plugin>
+                <groupId>com.diffplug.spotless</groupId>
+                <artifactId>spotless-maven-plugin</artifactId>
+                <version>2.44.0</version>
+                <configuration>
+                    <java>
+                        <googleJavaFormat />
+                        <importOrder />
+                        <removeUnusedImports />
+                    </java>
+                </configuration>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>check</goal>
+                            <goal>apply</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
 </project>
