@@ -1,29 +1,9 @@
-/**
- * 
- */
 package com.strandls.taxonomy;
 
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.lang.annotation.Annotation;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
-import java.util.Properties;
 import java.util.Set;
-import java.util.stream.Stream;
-
-import javax.ws.rs.core.Application;
 
 import org.glassfish.hk2.api.ServiceLocator;
-import org.glassfish.jersey.media.multipart.MultiPartFeature;
 import org.glassfish.jersey.server.spi.Container;
 import org.glassfish.jersey.server.spi.ContainerLifecycleListener;
 import org.glassfish.jersey.servlet.ServletContainer;
@@ -35,74 +15,58 @@ import org.slf4j.LoggerFactory;
 import com.google.inject.Injector;
 import com.strandls.authentication_utility.filter.InterceptorModule;
 
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiModel;
-import io.swagger.jaxrs.config.BeanConfig;
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.info.Info;
+import io.swagger.v3.oas.annotations.servers.Server;
+import jakarta.ws.rs.core.Application;
 
 /**
  * @author Abhishek Rudra
- *
  */
+@OpenAPIDefinition(info = @Info(title = "Taxonomy Module MicroServices", version = "1.0.0"), servers = {
+		@Server(url = "http://localhost:8080/taxonomy-api/api/") })
 public class ApplicationConfig extends Application {
 
 	private static final Logger logger = LoggerFactory.getLogger(ApplicationConfig.class);
 
-	/**
-	 * 
-	 */
 	public ApplicationConfig() {
-		InputStream in = Thread.currentThread().getContextClassLoader().getResourceAsStream("config.properties");
-
-		Properties properties = new Properties();
-		try {
-			properties.load(in);
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
-
-		BeanConfig beanConfig = new BeanConfig();
-		beanConfig.setVersion(properties.getProperty("version"));
-		beanConfig.setTitle(properties.getProperty("title"));
-		beanConfig.setSchemes(properties.getProperty("schemes").split(","));
-		beanConfig.setHost(properties.getProperty("host"));
-		beanConfig.setBasePath(properties.getProperty("basePath"));
-		beanConfig.setResourcePackage(properties.getProperty("resourcePackage"));
-		beanConfig.setPrettyPrint(new Boolean(properties.getProperty("prettyPrint")));
-		beanConfig.setScan(new Boolean(properties.getProperty("scan")));
+		logger.info("Initializing ApplicationConfig...");
 	}
 
 	@Override
 	public Set<Object> getSingletons() {
-
 		Set<Object> singletons = new HashSet<>();
-		singletons.add(new ContainerLifecycleListener() {
 
+		// Lifecycle listener to bridge Guice & HK2
+		singletons.add(new ContainerLifecycleListener() {
 			@Override
 			public void onStartup(Container container) {
+				logger.info("Starting up container and bridging Guice to HK2...");
+
 				ServletContainer servletContainer = (ServletContainer) container;
-				ServiceLocator serviceLocator = container.getApplicationHandler().getInjectionManager()
+				ServiceLocator locator = container.getApplicationHandler().getInjectionManager()
 						.getInstance(ServiceLocator.class);
-				GuiceBridge.getGuiceBridge().initializeGuiceBridge(serviceLocator);
-				GuiceIntoHK2Bridge guiceBridge = serviceLocator.getService(GuiceIntoHK2Bridge.class);
+
+				GuiceBridge.getGuiceBridge().initializeGuiceBridge(locator);
+				GuiceIntoHK2Bridge guiceBridge = locator.getService(GuiceIntoHK2Bridge.class);
+
 				Injector injector = (Injector) servletContainer.getServletContext()
 						.getAttribute(Injector.class.getName());
+
 				guiceBridge.bridgeGuiceInjector(injector);
 			}
 
 			@Override
 			public void onShutdown(Container container) {
-				/**
-				 * 
-				 */
+				logger.info("Container shutdown...");
 			}
 
 			@Override
 			public void onReload(Container container) {
-				/**
-				 * 
-				 */
+				logger.info("Container reload...");
 			}
 		});
+
 		singletons.add(new InterceptorModule());
 
 		return singletons;
@@ -110,64 +74,9 @@ public class ApplicationConfig extends Application {
 
 	@Override
 	public Set<Class<?>> getClasses() {
-		Set<Class<?>> resource = new HashSet<>();
-
-		try {
-			List<Class<?>> swaggerClass = getSwaggerAnnotationClassesFromPackage("com");
-			resource.addAll(swaggerClass);
-		} catch (ClassNotFoundException | URISyntaxException | IOException e) {
-			logger.error(e.getMessage());
-		}
-
-		resource.add(io.swagger.jaxrs.listing.SwaggerSerializers.class);
-		resource.add(io.swagger.jaxrs.listing.ApiListingResource.class);
-		
-		resource.add(MultiPartFeature.class);
-
-		return resource;
-	}
-
-	protected List<Class<?>> getSwaggerAnnotationClassesFromPackage(String packageName)
-			throws URISyntaxException, IOException, ClassNotFoundException {
-
-		List<String> classNames = getClassNamesFromPackage(packageName);
-		List<Class<?>> classes = new ArrayList<>();
-		for (String className : classNames) {
-			Class<?> cls = Class.forName(className);
-			Annotation[] annotations = cls.getAnnotations();
-
-			for (Annotation annotation : annotations) {
-				if (annotation instanceof Api || annotation instanceof ApiModel) {
-					classes.add(cls);
-				}
-			}
-		}
+		Set<Class<?>> classes = new HashSet<>();
+		classes.add(io.swagger.v3.jaxrs2.integration.resources.OpenApiResource.class);
 
 		return classes;
-	}
-
-	private static ArrayList<String> getClassNamesFromPackage(final String packageName)
-			throws URISyntaxException, IOException {
-
-		ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
-		ArrayList<String> names = new ArrayList<>();
-		URL packageURL = classLoader.getResource(packageName);
-
-		URI uri = new URI(packageURL.toString());
-		File folder = new File(uri.getPath());
-
-		try (Stream<Path> files = Files.find(Paths.get(folder.getAbsolutePath()), 999,
-				(p, bfa) -> bfa.isRegularFile())) {
-			files.forEach(file -> {
-				String name = file.toFile().getAbsolutePath()
-						.replaceAll(folder.getAbsolutePath() + File.separatorChar, "").replace(File.separatorChar, '.');
-				if (name.indexOf('.') != -1) {
-					name = packageName + '.' + name.substring(0, name.lastIndexOf('.'));
-					names.add(name);
-				}
-			});
-		}
-
-		return names;
 	}
 }
