@@ -1,20 +1,16 @@
-/**
- * 
- */
+/** */
 package com.strandls.taxonomy;
 
-import javax.ws.rs.core.HttpHeaders;
+import com.strandls.activity.controller.ActivityServiceApi;
 
-import com.strandls.activity.controller.ActivitySerivceApi;
+import jakarta.ws.rs.core.HttpHeaders;
 
 /**
  * @author Abhishek Rudra
- *
- * 
  */
 public class Headers {
 
-	public ActivitySerivceApi addActivityHeader(ActivitySerivceApi activityService, String authHeaders) {
+	public ActivityServiceApi addActivityHeader(ActivityServiceApi activityService, String authHeaders) {
 		activityService.getApiClient().addDefaultHeader(HttpHeaders.AUTHORIZATION, authHeaders);
 		return activityService;
 	}
