@@ -11,15 +11,15 @@ import org.slf4j.LoggerFactory;
 
 public class TaxonomyConfig {
 
-	private TaxonomyConfig() {}
+	private TaxonomyConfig() {
+	}
 
 	private static final Logger logger = LoggerFactory.getLogger(TaxonomyConfig.class);
 
 	private static Properties properties;
 
 	static {
-		InputStream in = Thread.currentThread().getContextClassLoader()
-				.getResourceAsStream("config.properties");
+		InputStream in = Thread.currentThread().getContextClassLoader().getResourceAsStream("config.properties");
 
 		properties = new Properties();
 		try {
@@ -28,18 +28,17 @@ public class TaxonomyConfig {
 			logger.error(e.getMessage());
 		}
 	}
-	
+
 	public static String getString(String key) {
 		return properties.getProperty(key);
 	}
-	
+
 	public static String fetchFileAsString(String fileName) throws IOException {
 		InputStream in = Thread.currentThread().getContextClassLoader().getResourceAsStream(fileName);
 		return IOUtils.toString(new InputStreamReader(in));
 	}
-	
+
 	public static int getInt(String key) {
 		return Integer.parseInt(getString(key));
 	}
-	
 }
