@@ -1,6 +1,3 @@
-/**
- * 
- */
 package com.strandls.taxonomy.controller;
 
 import java.io.IOException;
@@ -8,25 +5,6 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-
-import javax.inject.Inject;
-import javax.servlet.http.HttpServletRequest;
-import javax.transaction.Transactional;
-import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.DefaultValue;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.Status;
 
 import org.glassfish.jersey.media.multipart.FormDataBodyPart;
 import org.glassfish.jersey.media.multipart.FormDataMultiPart;
@@ -51,18 +29,36 @@ import com.strandls.taxonomy.service.TaxonomyDefinitionSerivce;
 import com.strandls.taxonomy.service.impl.TaxonomyESOperation;
 import com.strandls.taxonomy.util.TaxonomyUtil;
 
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.media.ArraySchema;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.parameters.RequestBody;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.inject.Inject;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.transaction.Transactional;
+import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.DELETE;
+import jakarta.ws.rs.DefaultValue;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.POST;
+import jakarta.ws.rs.PUT;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.QueryParam;
+import jakarta.ws.rs.WebApplicationException;
+import jakarta.ws.rs.core.Context;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.Response.Status;
 
-/**
- * @author Abhishek Rudra
- *
- */
-@Api("Taxonomy Services")
+@Tag(name = "Taxonomy Services", description = "APIs for taxonomy microservice")
 @Path(ApiConstants.V1 + ApiConstants.TAXONOMY)
+@Produces(MediaType.APPLICATION_JSON)
 public class TaxonomyDefinitionController {
 
 	@Inject
@@ -77,12 +73,11 @@ public class TaxonomyDefinitionController {
 	@GET
 	@Path("/{taxonomyConceptId}")
 	@Consumes(MediaType.TEXT_PLAIN)
-	@Produces(MediaType.APPLICATION_JSON)
-
-	@ApiOperation(value = "Find Taxonomy by ID", notes = "Returns Taxonomy details", response = TaxonomyDefinition.class)
-	@ApiResponses(value = { @ApiResponse(code = 404, message = "Taxonomy not found", response = String.class) })
-
-	public Response getTaxonomyConceptName(@PathParam("taxonomyConceptId") String taxonomyConceptId) {
+	@Operation(summary = "Find Taxonomy by ID", responses = {
+			@ApiResponse(responseCode = "200", description = "Taxonomy details", content = @Content(schema = @Schema(implementation = TaxonomyDefinition.class))),
+			@ApiResponse(responseCode = "404", description = "Taxonomy not found", content = @Content(schema = @Schema(implementation = String.class))) })
+	public Response getTaxonomyConceptName(
+			@Parameter(description = "Taxonomy Concept ID", required = true) @PathParam("taxonomyConceptId") String taxonomyConceptId) {
 		try {
 			Long id = Long.parseLong(taxonomyConceptId);
 			TaxonomyDefinition taxonomy = taxonomyService.fetchById(id);
@@ -95,12 +90,11 @@ public class TaxonomyDefinitionController {
 	@GET
 	@Path("/show/{taxonId}")
 	@Consumes(MediaType.TEXT_PLAIN)
-	@Produces(MediaType.APPLICATION_JSON)
-
-	@ApiOperation(value = "Find Taxonomy Details by ID", notes = "Returns Taxonomy details", response = TaxonomyDefinitionShow.class)
-	@ApiResponses(value = { @ApiResponse(code = 404, message = "Taxonomy not found", response = String.class) })
-
-	public Response getTaxonomyDetails(@PathParam("taxonId") String taxonId) {
+	@Operation(summary = "Find Taxonomy Details by ID", responses = {
+			@ApiResponse(responseCode = "200", description = "Taxonomy details", content = @Content(schema = @Schema(implementation = TaxonomyDefinitionShow.class))),
+			@ApiResponse(responseCode = "404", description = "Taxonomy not found", content = @Content(schema = @Schema(implementation = String.class))) })
+	public Response getTaxonomyDetails(
+			@Parameter(description = "Taxon ID", required = true) @PathParam("taxonId") String taxonId) {
 		try {
 			Long id = Long.parseLong(taxonId);
 			TaxonomyDefinitionShow taxonomyDefinitionShow = taxonomyService.getTaxonomyDetails(id);
@@ -113,15 +107,17 @@ public class TaxonomyDefinitionController {
 	@GET
 	@Path("/namelist")
 	@Consumes(MediaType.TEXT_PLAIN)
-	@Produces(MediaType.APPLICATION_JSON)
-
-	@ApiOperation(value = "Get taxonomy name list ", notes = "Returns taxonomy details", response = TaxonomyNameListResponse.class)
-	@ApiResponses(value = { @ApiResponse(code = 404, message = "Taxonomy list not found", response = String.class) })
-	public Response getTaxonomyNameList(@Context HttpServletRequest request, @QueryParam("taxonId") Long taxonId,
-			@QueryParam("classificationId") Long classificationId, @QueryParam("rankList") String rankList,
-			@QueryParam("statusList") String statusList, @QueryParam("positionList") String positionList,
-			@DefaultValue("-1") @QueryParam("limit") Integer limit,
-			@DefaultValue("-1") @QueryParam("offset") Integer offset) {
+	@Operation(summary = "Get taxonomy name list", responses = {
+			@ApiResponse(responseCode = "200", description = "Taxonomy Name List", content = @Content(schema = @Schema(implementation = TaxonomyNameListResponse.class))),
+			@ApiResponse(responseCode = "404", description = "Taxonomy list not found", content = @Content(schema = @Schema(implementation = String.class))) })
+	public Response getTaxonomyNameList(@Context HttpServletRequest request,
+			@Parameter(description = "Taxon ID") @QueryParam("taxonId") Long taxonId,
+			@Parameter(description = "Classification ID") @QueryParam("classificationId") Long classificationId,
+			@Parameter(description = "Comma separated rank list") @QueryParam("rankList") String rankList,
+			@Parameter(description = "Status list") @QueryParam("statusList") String statusList,
+			@Parameter(description = "Position list") @QueryParam("positionList") String positionList,
+			@Parameter(description = "Limit", example = "-1") @DefaultValue("-1") @QueryParam("limit") Integer limit,
+			@Parameter(description = "Offset", example = "-1") @DefaultValue("-1") @QueryParam("offset") Integer offset) {
 		try {
 			TaxonomyNameListResponse response = taxonomyService.getTaxonomyNameList(taxonId, classificationId, rankList,
 					statusList, positionList, limit, offset);
@@ -132,13 +128,13 @@ public class TaxonomyDefinitionController {
 		}
 	}
 
-	@Path("upload")
 	@POST
+	@Path("upload")
 	@Consumes({ MediaType.MULTIPART_FORM_DATA })
-	@Produces(MediaType.APPLICATION_JSON)
-	@ApiOperation(value = "Upload the file for taxon definition", notes = "Returns succuess failure", response = FileMetadata.class)
-	@ApiResponses(value = { @ApiResponse(code = 400, message = "file not present", response = String.class),
-			@ApiResponse(code = 500, message = "ERROR", response = String.class) })
+	@Operation(summary = "Upload the file for taxon definition", requestBody = @RequestBody(required = true, description = "Multi-part form data with file and associated info"), responses = {
+			@ApiResponse(responseCode = "200", description = "Success/failure", content = @Content(schema = @Schema(implementation = FileMetadata.class))),
+			@ApiResponse(responseCode = "400", description = "file not present", content = @Content(schema = @Schema(implementation = String.class))),
+			@ApiResponse(responseCode = "500", description = "ERROR", content = @Content(schema = @Schema(implementation = String.class))) })
 	@ValidateUser
 	public Response upload(@Context HttpServletRequest request, final FormDataMultiPart multiPart) {
 		try {
@@ -150,26 +146,24 @@ public class TaxonomyDefinitionController {
 		}
 	}
 
-	@Path(ApiConstants.UPLOAD + ApiConstants.SEARCH)
 	@POST
+	@Path(ApiConstants.UPLOAD + ApiConstants.SEARCH)
 	@Consumes({ MediaType.MULTIPART_FORM_DATA })
-	@Produces(MediaType.APPLICATION_JSON)
-	@ApiOperation(value = "Upload the file for taxon definition", notes = "Returns succuess failure", response = Map.class)
-	@ApiResponses(value = { @ApiResponse(code = 400, message = "file not present", response = String.class),
-			@ApiResponse(code = 500, message = "ERROR", response = String.class) })
+	@Operation(summary = "Upload the file for taxon definition search", requestBody = @RequestBody(required = true, description = "Multi-part file for name matching"), responses = {
+			@ApiResponse(responseCode = "200", description = "Success/failure as a map", content = @Content(schema = @Schema(implementation = Map.class))),
+			@ApiResponse(responseCode = "400", description = "file not present", content = @Content(schema = @Schema(implementation = String.class))),
+			@ApiResponse(responseCode = "500", description = "ERROR", content = @Content(schema = @Schema(implementation = String.class))) })
 	public Response uploadSearch(final FormDataMultiPart multiPart) {
 		FormDataBodyPart filePart = multiPart.getField("file");
 		Integer index = Integer.valueOf(multiPart.getField("column").getValue());
 		if (filePart == null) {
 			return Response.status(Response.Status.BAD_REQUEST).entity("File not present").build();
-		} else {
-			Map<String, Object> result;
-			try {
-				result = taxonomyService.nameMatching(filePart, index);
-				return Response.ok().entity(result).build();
-			} catch (IOException e) {
-				return Response.status(Status.BAD_REQUEST).entity(e.getMessage()).build();
-			}
+		}
+		try {
+			Map<String, Object> result = taxonomyService.nameMatching(filePart, index);
+			return Response.ok().entity(result).build();
+		} catch (IOException e) {
+			return Response.status(Status.BAD_REQUEST).entity(e.getMessage()).build();
 		}
 	}
 
@@ -177,11 +171,8 @@ public class TaxonomyDefinitionController {
 	@Path("list")
 	@Produces(MediaType.APPLICATION_JSON)
 	@ValidateUser
-	@ApiOperation(value = "save the taxonomy list", notes = "return the saved taxonomy", response = TaxonomyDefinition.class, responseContainer = "List")
-	@ApiResponses(value = {
-			@ApiResponse(code = 400, message = "failed to save the taxon definition", response = String.class) })
-	public Response saveTaxonomyList(@Context HttpServletRequest request,
-			@ApiParam("taxonomyList") List<TaxonomySave> taxonomyList) {
+	@Operation(summary = "Save the taxonomy list", requestBody = @RequestBody(required = true, content = @Content(array = @ArraySchema(schema = @Schema(implementation = TaxonomySave.class)))), responses = @ApiResponse(responseCode = "200", description = "Saved list", content = @Content(array = @ArraySchema(schema = @Schema(implementation = TaxonomyDefinition.class)))))
+	public Response saveTaxonomyList(@Context HttpServletRequest request, List<TaxonomySave> taxonomyList) {
 		try {
 			List<TaxonomyDefinition> taxonomyDefinition = taxonomyService.saveList(request, taxonomyList);
 			return Response.status(Status.OK).entity(taxonomyDefinition).build();
@@ -193,11 +184,8 @@ public class TaxonomyDefinitionController {
 	@POST
 	@Produces(MediaType.APPLICATION_JSON)
 	@ValidateUser
-	@ApiOperation(value = "save the taxonomy", notes = "return the saved taxonomy", response = TaxonomyDefinition.class)
-	@ApiResponses(value = {
-			@ApiResponse(code = 400, message = "failed to save the taxon definition", response = String.class) })
-	public Response saveTaxonomy(@Context HttpServletRequest request,
-			@ApiParam("taxonSave") TaxonomySave taxonomySave) {
+	@Operation(summary = "Save the taxonomy", requestBody = @RequestBody(required = true, content = @Content(schema = @Schema(implementation = TaxonomySave.class))), responses = @ApiResponse(responseCode = "200", description = "Saved Taxonomy", content = @Content(schema = @Schema(implementation = TaxonomyDefinition.class))))
+	public Response saveTaxonomy(@Context HttpServletRequest request, TaxonomySave taxonomySave) {
 		try {
 			TaxonomyDefinition taxonomyDefinition = taxonomyService.save(request, taxonomySave);
 			return Response.status(Status.OK).entity(taxonomyDefinition).build();
@@ -209,11 +197,10 @@ public class TaxonomyDefinitionController {
 	@GET
 	@Path("/nameSearch")
 	@Produces(MediaType.APPLICATION_JSON)
-	@ApiOperation(value = "Get taxonomy based on the canonical name and rank", notes = "return the found taxonomy", response = TaxonomySearch.class)
-	@ApiResponses(value = {
-			@ApiResponse(code = 400, message = "failed to get the taxon definition", response = String.class) })
-	public Response getByNameSearch(@QueryParam("scientificName") String scientificName,
-			@QueryParam("rankName") String rankName) {
+	@Operation(summary = "Get taxonomy based on the canonical name and rank", responses = @ApiResponse(responseCode = "200", description = "Found taxonomy", content = @Content(schema = @Schema(implementation = TaxonomySearch.class))))
+	public Response getByNameSearch(
+			@Parameter(description = "Scientific Name") @QueryParam("scientificName") String scientificName,
+			@Parameter(description = "Rank name") @QueryParam("rankName") String rankName) {
 		try {
 			TaxonomySearch taxonomySearch = taxonomyService.getByNameSearch(scientificName, rankName);
 			return Response.status(Status.OK).entity(taxonomySearch).build();
@@ -227,10 +214,8 @@ public class TaxonomyDefinitionController {
 	@Path("/search")
 	@Produces(MediaType.APPLICATION_JSON)
 	@Transactional
-	@ApiOperation(value = "Search taxonomy based on the name", notes = "return the found taxonomy", response = Object.class)
-	@ApiResponses(value = {
-			@ApiResponse(code = 400, message = "failed to get the taxon definition", response = String.class) })
-	public Response search(@QueryParam("term") String term) {
+	@Operation(summary = "Search taxonomy based on the name", responses = @ApiResponse(responseCode = "200", description = "Found taxonomy search result", content = @Content(schema = @Schema(implementation = Object.class))))
+	public Response search(@Parameter(description = "Search term") @QueryParam("term") String term) {
 		try {
 			Object name = taxonomyDefinitionDao.search(term);
 			return Response.status(Status.OK).entity(name).build();
@@ -238,18 +223,16 @@ public class TaxonomyDefinitionController {
 			throw new WebApplicationException(
 					Response.status(Response.Status.INTERNAL_SERVER_ERROR).entity(e.getMessage()).build());
 		}
-
 	}
 
 	@GET
 	@Path("/retrieve/specificSearch")
 	@Produces(MediaType.APPLICATION_JSON)
 	@Transactional
-	@ApiOperation(value = "Search taxonomy based on the Ids", notes = "return the found taxonomy", response = Object.class)
-	@ApiResponses(value = {
-			@ApiResponse(code = 400, message = "failed to get the taxon definition", response = String.class) })
-	public Response specificSearch(@QueryParam("term") String term, @QueryParam("classification") Long classificationId,
-			@QueryParam("taxonid") Long taxonid) {
+	@Operation(summary = "Search taxonomy based on the Ids", responses = @ApiResponse(responseCode = "200", description = "Found taxonomy", content = @Content(array = @ArraySchema(schema = @Schema(implementation = String.class)))))
+	public Response specificSearch(@Parameter(description = "Search term") @QueryParam("term") String term,
+			@Parameter(description = "Classification ID") @QueryParam("classification") Long classificationId,
+			@Parameter(description = "Taxon ID") @QueryParam("taxonid") Long taxonid) {
 		try {
 			List<String> resultTaxonIds = taxonomyDefinitionDao.specificSearch(term, taxonid);
 			return Response.status(Status.OK).entity(resultTaxonIds).build();
@@ -262,12 +245,11 @@ public class TaxonomyDefinitionController {
 	@GET
 	@Path(ApiConstants.NAMES + "/{taxonomyId}")
 	@Consumes(MediaType.TEXT_PLAIN)
-	@Produces(MediaType.APPLICATION_JSON)
-
-	@ApiOperation(value = "get the common name and synonyms", notes = "return taxonoicNames based on taxonomyId", response = TaxonomicNames.class)
-	@ApiResponses(value = { @ApiResponse(code = 400, message = "unable to get the names", response = String.class) })
-
-	public Response getNames(@PathParam("taxonomyId") String taxonomyId) {
+	@Operation(summary = "Get the common name and synonyms", responses = {
+			@ApiResponse(responseCode = "200", description = "Taxonomic names based on taxonomyId", content = @Content(schema = @Schema(implementation = TaxonomicNames.class))),
+			@ApiResponse(responseCode = "400", description = "unable to get the names", content = @Content(schema = @Schema(implementation = String.class))) })
+	public Response getNames(
+			@Parameter(description = "Taxonomy ID", required = true) @PathParam("taxonomyId") String taxonomyId) {
 		try {
 			Long taxonId = Long.parseLong(taxonomyId);
 			TaxonomicNames result = taxonomyService.findSynonymCommonName(taxonId);
@@ -280,15 +262,11 @@ public class TaxonomyDefinitionController {
 	@PUT
 	@Path("name")
 	@Consumes(MediaType.TEXT_PLAIN)
-	@Produces(MediaType.APPLICATION_JSON)
-
+	@Operation(summary = "Update the name of taxonomy", requestBody = @RequestBody(required = true, description = "The new scientific name"), responses = @ApiResponse(responseCode = "200", description = "Updated taxonomy", content = @Content(schema = @Schema(implementation = TaxonomyDefinitionShow.class))))
 	@ValidateUser
-
-	@ApiOperation(value = "Update the name of taxonomy", notes = "Update the name. input name should be scientific name", response = TaxonomyDefinitionShow.class)
-	@ApiResponses(value = {
-			@ApiResponse(code = 400, message = "failed to update the name of taxonomy definition", response = String.class) })
-	public Response updateName(@Context HttpServletRequest request, @QueryParam("taxonId") Long taxonId,
-			@QueryParam("taxonName") String taxonName) {
+	public Response updateName(@Context HttpServletRequest request,
+			@Parameter(description = "Taxon ID") @QueryParam("taxonId") Long taxonId,
+			@Parameter(description = "Taxon Name") @QueryParam("taxonName") String taxonName) {
 		try {
 			TaxonomyDefinitionShow taxonomyDefinitionShow = taxonomyService.updateName(request, taxonId, taxonName);
 			return Response.status(Status.OK).entity(taxonomyDefinitionShow).build();
@@ -300,15 +278,9 @@ public class TaxonomyDefinitionController {
 
 	@PUT
 	@Path("status")
-	@Produces(MediaType.APPLICATION_JSON)
-
+	@Operation(summary = "Update the taxonomy status (accepted or synonym)", requestBody = @RequestBody(required = true, content = @Content(schema = @Schema(implementation = TaxonomyStatusUpdate.class))), responses = @ApiResponse(responseCode = "200", description = "Updated taxonomy status", content = @Content(schema = @Schema(implementation = TaxonomyDefinitionShow.class))))
 	@ValidateUser
-
-	@ApiOperation(value = "Update the name of taxonomy", notes = "Update the status. Status should be either accepted or synonym", response = TaxonomyDefinitionShow.class)
-	@ApiResponses(value = {
-			@ApiResponse(code = 400, message = "failed to update the name of taxonomy definition", response = String.class) })
-	public Response updateStatus(@Context HttpServletRequest request,
-			@ApiParam("status") TaxonomyStatusUpdate taxonomyStatusUpdate) {
+	public Response updateStatus(@Context HttpServletRequest request, TaxonomyStatusUpdate taxonomyStatusUpdate) {
 		try {
 			TaxonomyDefinitionShow taxonomyDefinitionShow = taxonomyService.updateStatus(request, taxonomyStatusUpdate);
 			return Response.status(Status.OK).entity(taxonomyDefinitionShow).build();
@@ -320,15 +292,9 @@ public class TaxonomyDefinitionController {
 
 	@PUT
 	@Path("position")
-	@Produces(MediaType.APPLICATION_JSON)
-
+	@Operation(summary = "Update the position of taxonomy (raw or working)", requestBody = @RequestBody(required = true, content = @Content(schema = @Schema(implementation = TaxonomyPositionUpdate.class))), responses = @ApiResponse(responseCode = "200", description = "Updated taxonomy position", content = @Content(schema = @Schema(implementation = TaxonomyDefinitionShow.class))))
 	@ValidateUser
-
-	@ApiOperation(value = "Update the position of taxonomy", notes = "Update the position. Position should be raw or working", response = TaxonomyDefinition.class)
-	@ApiResponses(value = {
-			@ApiResponse(code = 400, message = "failed to update the position of taxonomy definition", response = String.class) })
-	public Response updatePosition(@Context HttpServletRequest request,
-			@ApiParam("status") TaxonomyPositionUpdate taxonomyPositionUpdate) {
+	public Response updatePosition(@Context HttpServletRequest request, TaxonomyPositionUpdate taxonomyPositionUpdate) {
 		try {
 			TaxonomyDefinitionShow taxonomyDefinitionShow = taxonomyService.updatePosition(request,
 					taxonomyPositionUpdate);
@@ -342,19 +308,14 @@ public class TaxonomyDefinitionController {
 	@POST
 	@Path(ApiConstants.UPDATE + ApiConstants.SYNONYM + "/{taxonId}")
 	@Consumes(MediaType.APPLICATION_JSON)
-	@Produces(MediaType.APPLICATION_JSON)
-
 	@ValidateUser
-
-	@ApiOperation(value = "update and add synonyms", notes = "return synonyms based on taxonomyId", response = TaxonomyDefinition.class, responseContainer = "List")
-	@ApiResponses(value = { @ApiResponse(code = 400, message = "unable to add the names", response = String.class) })
-
-	public Response updateAddSynonym(@Context HttpServletRequest request, @QueryParam("speciesId") String speciesId,
-			@PathParam("taxonId") String taxonId, @ApiParam(name = "synonymData") SynonymData synonymData) {
+	@Operation(summary = "Update and add synonyms to taxonomy", requestBody = @RequestBody(required = true, content = @Content(schema = @Schema(implementation = SynonymData.class))), responses = @ApiResponse(responseCode = "200", description = "Updated synonyms", content = @Content(array = @ArraySchema(schema = @Schema(implementation = TaxonomyDefinition.class)))))
+	public Response updateAddSynonym(@Context HttpServletRequest request,
+			@Parameter(description = "Species ID") @QueryParam("speciesId") String speciesId,
+			@Parameter(description = "Taxon ID", required = true) @PathParam("taxonId") String taxonId,
+			SynonymData synonymData) {
 		try {
-			Long sId = null;
-			if (speciesId != null)
-				sId = Long.parseLong(speciesId);
+			Long sId = speciesId != null ? Long.parseLong(speciesId) : null;
 			Long tId = Long.parseLong(taxonId);
 			List<TaxonomyDefinition> result = taxonomyService.updateAddSynonym(request, sId, tId, synonymData);
 			return Response.status(Status.OK).entity(result).build();
@@ -366,19 +327,14 @@ public class TaxonomyDefinitionController {
 	@DELETE
 	@Path(ApiConstants.REMOVE + ApiConstants.SYNONYM + "/{taxonId}/{synonymId}")
 	@Consumes(MediaType.TEXT_PLAIN)
-	@Produces(MediaType.APPLICATION_JSON)
-
 	@ValidateUser
-
-	@ApiOperation(value = "delete synonyms", notes = "return list of avaible synonyms", response = TaxonomyDefinition.class, responseContainer = "List")
-	@ApiResponses(value = { @ApiResponse(code = 400, message = "unable to delete the names", response = String.class) })
-
-	public Response removeSynonyms(@Context HttpServletRequest request, @QueryParam("speciesId") String speciesId,
-			@PathParam("taxonId") String taxonId, @PathParam("synonymId") String synonymId) {
+	@Operation(summary = "Delete synonym for a taxonomy", responses = @ApiResponse(responseCode = "200", description = "List of available synonyms", content = @Content(array = @ArraySchema(schema = @Schema(implementation = TaxonomyDefinition.class)))))
+	public Response removeSynonyms(@Context HttpServletRequest request,
+			@Parameter(description = "Species ID") @QueryParam("speciesId") String speciesId,
+			@Parameter(description = "Taxon ID", required = true) @PathParam("taxonId") String taxonId,
+			@Parameter(description = "Synonym ID", required = true) @PathParam("synonymId") String synonymId) {
 		try {
-			Long sId = null;
-			if (speciesId != null)
-				sId = Long.parseLong(speciesId);
+			Long sId = speciesId != null ? Long.parseLong(speciesId) : null;
 			Long tId = Long.parseLong(taxonId);
 			Long synonId = Long.parseLong(synonymId);
 			List<TaxonomyDefinition> result = taxonomyService.deleteSynonym(request, sId, tId, synonId);
@@ -391,24 +347,20 @@ public class TaxonomyDefinitionController {
 	@PUT
 	@Path(ApiConstants.ELASTIC + ApiConstants.UPDATE)
 	@Consumes(MediaType.TEXT_PLAIN)
-	@Produces(MediaType.APPLICATION_JSON)
-
 	@ValidateUser
-
-	@ApiOperation(value = "Update the elastic index for taxon", notes = "Re-index the elastic for given taxon Ids", response = TaxonomyDefinition.class, responseContainer = "List")
-	@ApiResponses(value = {
-			@ApiResponse(code = 400, message = "unable to reIndex the taxon Ids", response = String.class) })
-
-	public Response updateElastic(@Context HttpServletRequest request, @QueryParam("taxonIds") String taxonIdsString) {
+	@Operation(summary = "Update the elastic index for taxon", responses = @ApiResponse(responseCode = "200", description = "Elastic index updated", content = @Content(array = @ArraySchema(schema = @Schema(implementation = MapQueryResponse.class)))))
+	public Response updateElastic(@Context HttpServletRequest request,
+			@Parameter(description = "Comma separated Taxon IDs") @QueryParam("taxonIds") String taxonIdsString) {
 		try {
 			if (TaxonomyUtil.isAdmin(request)) {
 				List<Long> taxonIds = Arrays.asList(taxonIdsString.split(",")).stream().map(Long::parseLong)
 						.collect(Collectors.toList());
 				List<MapQueryResponse> mapQueryResponses = taxonomyESOperation.pushToElastic(taxonIds);
 				return Response.status(Status.OK).entity(mapQueryResponses).build();
-			} else
+			} else {
 				throw new WebApplicationException(
 						Response.status(Response.Status.UNAUTHORIZED).entity("Only admin can do the reindex").build());
+			}
 		} catch (Exception e) {
 			throw new WebApplicationException(
 					Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build());
@@ -418,21 +370,17 @@ public class TaxonomyDefinitionController {
 	@PUT
 	@Path(ApiConstants.ELASTIC + ApiConstants.REINDEX)
 	@Consumes(MediaType.TEXT_PLAIN)
-	@Produces(MediaType.APPLICATION_JSON)
-
 	@ValidateUser
-
-	@ApiOperation(value = "Update the elastic index for taxon", notes = "Re-index the complete elastic for all taxon Ids", response = TaxonomyDefinition.class, responseContainer = "List")
-	@ApiResponses(value = { @ApiResponse(code = 400, message = "unable to delete the names", response = String.class) })
-
+	@Operation(summary = "Reindex the complete elastic for all taxon Ids", responses = @ApiResponse(responseCode = "200", description = "Elastic index reindexed", content = @Content(array = @ArraySchema(schema = @Schema(implementation = MapQueryResponse.class)))))
 	public Response reIndexElastic(@Context HttpServletRequest request) {
 		try {
 			if (TaxonomyUtil.isAdmin(request)) {
 				List<MapQueryResponse> mapQueryResponses = taxonomyESOperation.reIndexElastic();
 				return Response.status(Status.OK).entity(mapQueryResponses).build();
-			} else
+			} else {
 				throw new WebApplicationException(
 						Response.status(Response.Status.UNAUTHORIZED).entity("Only admin can do the reindex").build());
+			}
 		} catch (Exception e) {
 			throw new WebApplicationException(
 					Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build());
@@ -442,15 +390,9 @@ public class TaxonomyDefinitionController {
 	@POST
 	@Path(ApiConstants.ADD + ApiConstants.COMMENT)
 	@Consumes(MediaType.APPLICATION_JSON)
-	@Produces(MediaType.APPLICATION_JSON)
-
 	@ValidateUser
-
-	@ApiOperation(value = "log taxonomy Comment", notes = "Return the logged the comment", response = Activity.class)
-	@ApiResponses(value = { @ApiResponse(code = 400, message = "unable to log comment", response = String.class) })
-
-	public Response addComment(@Context HttpServletRequest request,
-			@ApiParam(name = "loggingData") CommentLoggingData loggingData) {
+	@Operation(summary = "Log taxonomy Comment", requestBody = @RequestBody(required = true, content = @Content(schema = @Schema(implementation = CommentLoggingData.class))), responses = @ApiResponse(responseCode = "200", description = "Logged comment", content = @Content(schema = @Schema(implementation = Activity.class))))
+	public Response addComment(@Context HttpServletRequest request, CommentLoggingData loggingData) {
 		try {
 			Activity result = taxonomyService.logComment(request, loggingData);
 			return Response.status(Status.OK).entity(result).build();
@@ -461,33 +403,27 @@ public class TaxonomyDefinitionController {
 
 	/**
 	 * Only for migration purpose
-	 * 
+	 *
 	 * @param request
 	 * @return
 	 */
 	@PUT
 	@Path(ApiConstants.NAMES + ApiConstants.ITALICISED)
 	@Consumes(MediaType.TEXT_PLAIN)
-	@Produces(MediaType.APPLICATION_JSON)
-
 	@ValidateUser
-
-	@ApiOperation(value = "Update Italicised form for all the taxonomy definition", response = Map.class)
-	@ApiResponses(value = {
-			@ApiResponse(code = 400, message = "unable to upate the italicised form for the names", response = String.class) })
-
+	@Operation(summary = "Update Italicised form for all the taxonomy definition", responses = @ApiResponse(responseCode = "200", description = "Map of taxonomy definitions", content = @Content(schema = @Schema(implementation = Map.class))))
 	public Response updateItalicisedForm(@Context HttpServletRequest request) {
 		try {
 			if (TaxonomyUtil.isAdmin(request)) {
 				Map<String, TaxonomyDefinition> mapResponse = taxonomyService.updateItalicisedForm();
 				return Response.status(Status.OK).entity(mapResponse).build();
-			} else
+			} else {
 				throw new WebApplicationException(Response.status(Response.Status.UNAUTHORIZED)
 						.entity("Only admin can do the complete update of the name").build());
+			}
 		} catch (Exception e) {
 			throw new WebApplicationException(
 					Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build());
 		}
 	}
-
 }
