@@ -1,11 +1,7 @@
-/**
- * 
- */
+/** */
 package com.strandls.taxonomy.dao;
 
 import java.util.List;
-
-import javax.inject.Inject;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
@@ -17,17 +13,18 @@ import org.slf4j.LoggerFactory;
 import com.strandls.taxonomy.pojo.AcceptedSynonym;
 import com.strandls.taxonomy.util.AbstractDAO;
 
+import jakarta.inject.Inject;
+
 /**
  * @author Abhishek Rudra
- *
  */
 public class AcceptedSynonymDao extends AbstractDAO<AcceptedSynonym, Long> {
 
 	private final Logger logger = LoggerFactory.getLogger(AcceptedSynonymDao.class);
 
 	private static final String ACCEPTED_ID = "acceptedId";
-	private static final String SYNONYM_ID= "synonymId";
-	
+	private static final String SYNONYM_ID = "synonymId";
+
 	/**
 	 * @param sessionFactory
 	 */
@@ -49,7 +46,7 @@ public class AcceptedSynonymDao extends AbstractDAO<AcceptedSynonym, Long> {
 		}
 		return result;
 	}
-	
+
 	public AcceptedSynonym createAcceptedSynonym(Long acceptedId, Long synonymId) {
 		AcceptedSynonym acceptedSynonym = findByAccpetedIdSynonymId(acceptedId, synonymId);
 		if (acceptedSynonym == null) {
@@ -98,7 +95,7 @@ public class AcceptedSynonymDao extends AbstractDAO<AcceptedSynonym, Long> {
 		}
 		return result;
 	}
-	
+
 	@SuppressWarnings("unchecked")
 	public List<AcceptedSynonym> findByAccepetdId(Long acceptedId) {
 		String qry = "from AcceptedSynonym where acceptedId = :acceptedId";
@@ -119,6 +116,7 @@ public class AcceptedSynonymDao extends AbstractDAO<AcceptedSynonym, Long> {
 
 	/**
 	 * Transfer all the synonym from one accepted name to another
+	 *
 	 * @param taxonId
 	 * @param newTaxonId
 	 */
