@@ -1,7 +1,4 @@
-
-/**
- * 
- */
+/** */
 package com.strandls.taxonomy.dao;
 
 import com.google.inject.AbstractModule;
@@ -9,7 +6,6 @@ import com.google.inject.Scopes;
 
 /**
  * @author Abhishek Rudra
- *
  */
 public class TaxonomyDaoModule extends AbstractModule {
 
