@@ -1,6 +1,4 @@
-/**
- * 
- */
+/** */
 package com.strandls.taxonomy.dao;
 
 import java.sql.Timestamp;
@@ -9,8 +7,6 @@ import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 import java.util.Set;
-
-import javax.inject.Inject;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
@@ -24,12 +20,14 @@ import com.strandls.taxonomy.TaxonomyConfig;
 import com.strandls.taxonomy.pojo.TaxonomyDefinition;
 import com.strandls.taxonomy.pojo.TaxonomyRegistry;
 import com.strandls.taxonomy.pojo.response.TaxonRelation;
+import com.strandls.taxonomy.pojo.response.TaxonomyNamelistItem;
 import com.strandls.taxonomy.pojo.response.TaxonomyRegistryResponse;
 import com.strandls.taxonomy.util.AbstractDAO;
 
+import jakarta.inject.Inject;
+
 /**
  * @author Abhishek Rudra
- *
  */
 public class TaxonomyRegistryDao extends AbstractDAO<TaxonomyRegistry, Long> {
 
@@ -51,7 +49,6 @@ public class TaxonomyRegistryDao extends AbstractDAO<TaxonomyRegistry, Long> {
 	@Inject
 	protected TaxonomyRegistryDao(SessionFactory sessionFactory) {
 		super(sessionFactory);
-
 	}
 
 	@Override
@@ -155,7 +152,7 @@ public class TaxonomyRegistryDao extends AbstractDAO<TaxonomyRegistry, Long> {
 					+ " from taxonomy_definition as t, taxonomy_registry as tR"
 					+ " where t.id=tR.taxon_definition_id and t.is_deleted=false  and tR.classification_id=:classification_id and "
 					+ " nlevel(tR.path) = 2" + " order by nlevel(tR.path), t.name";
-			query = session.createNativeQuery(queryString).setResultSetMapping("TaxonomyRelation");
+			query = session.createNativeQuery(queryString, TaxonomyNamelistItem.class);
 		} else {
 			String parentCheck = " ";
 			if (expandTaxon && taxonIds != null && !taxonIds.isEmpty()) {
@@ -173,7 +170,7 @@ public class TaxonomyRegistryDao extends AbstractDAO<TaxonomyRegistry, Long> {
 					+ " from taxonomy_definition as t, taxonomy_registry as tR"
 					+ " where t.id=tR.taxon_definition_id and t.is_deleted=false  and tR.classification_id=:classification_id and "
 					+ " tR.path ~ lquery(:parentCheck) and nlevel(tR.path) > 1 " + " order by nlevel(tR.path), t.name";
-			query = session.createNativeQuery(queryString).setResultSetMapping("TaxonomyRelation");
+			query = session.createNativeQuery(queryString, TaxonomyNamelistItem.class);
 			query.setParameter(PARENT_CHECK, parentCheck);
 		}
 		classificationId = classificationId == null ? CLASSIFICATION_ID : classificationId;
@@ -197,7 +194,7 @@ public class TaxonomyRegistryDao extends AbstractDAO<TaxonomyRegistry, Long> {
 			String sqlString = "select cast(taxon_definition_id as varchar) from taxonomy_registry where path @> "
 					+ "any(select path from taxonomy_registry where taxon_definition_id in (:taxonIds) and classification_id=:classificationId) and "
 					+ "classification_id=:classificationId";
-			if(taxonIds==null) {
+			if (taxonIds == null) {
 				sqlString = "select cast(taxon_definition_id as varchar) from taxonomy_registry where path @> "
 						+ "any(select path from taxonomy_registry where classification_id=:classificationId) and "
 						+ "classification_id=:classificationId";
@@ -242,12 +239,11 @@ public class TaxonomyRegistryDao extends AbstractDAO<TaxonomyRegistry, Long> {
 
 	/**
 	 * Code below this point is only for the migration purpose.
-	 * 
+	 *
 	 * @param taxonId
 	 * @param classificationIds
 	 * @return
 	 */
-
 	@SuppressWarnings({ "unchecked", "rawtypes" })
 	public List<TaxonomyRegistry> getOldHierarchy(Long taxonId, List<Long> classificationIds) {
 		Session session = sessionFactory.openSession();
@@ -311,27 +307,30 @@ public class TaxonomyRegistryDao extends AbstractDAO<TaxonomyRegistry, Long> {
 	}
 
 	/**
-	 * 
 	 * Order for choosing candidate is as follow 821 - Catalog of Life 820 -
 	 * FishBase Taxonomy Hierarchy 819 - IUCN Taxonomy Hierarchy (2010) 818 - GBIF
 	 * Taxonomy Hierarchy 817 - Author Contributed Taxonomy Hierarchy 265798 -
 	 * Combined Taxonomy Hierarchy
-	 * 
+	 *
+	 * <p>
 	 * If we found any duplicate candidate the we fall back to IBP hierarchy. so we
 	 * are maintaining both the record Pulling out one of the hierarchy mentioned
 	 * above in given order and IBP hierarchy
-	 * 
+	 *
+	 * <p>
 	 * 265799 - IBP hierarchy
-	 * 
+	 *
+	 * <p>
 	 * Query to execute is as follow
-	 * 
+	 *
+	 * <p>
 	 * select * from taxonomy_registry_backup where taxon_definition_id = :taxonId
 	 * and (classification_id = :classificationId or classification_id = ( select
 	 * classification_id from taxonomy_registry_backup where taxon_definition_id =
 	 * :taxonId group by classification_id order by case classification_id when 821
 	 * then 1 when 820 then 2 when 819 then 3 when 818 then 4 when 817 then 5 when
 	 * 265798 then 6 end limit 1 )) order by classification_id desc
-	 * 
+	 *
 	 * @param taxonId
 	 * @return
 	 */
@@ -408,5 +407,4 @@ public class TaxonomyRegistryDao extends AbstractDAO<TaxonomyRegistry, Long> {
 			session.close();
 		}
 	}
-
 }
