@@ -2,24 +2,21 @@ package com.strandls.taxonomy.pojo;
 
 import java.util.Date;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.SequenceGenerator;
-import javax.persistence.Table;
-import javax.persistence.Transient;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.strandls.utility.pojo.Language;
 
-/**
- * 
- * @author vilay
- *
- */
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.SequenceGenerator;
+import jakarta.persistence.Table;
+import jakarta.persistence.Transient;
 
+/**
+ * @author vilay
+ */
 @Entity
 @Table(name = "common_names")
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -169,5 +166,4 @@ public class CommonName {
 	public void setLanguage(Language language) {
 		this.language = language;
 	}
-
 }
