@@ -1,31 +1,26 @@
-/**
- * 
- */
+/** */
 package com.strandls.taxonomy.pojo;
 
 import java.io.Serializable;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 
 /**
  * @author Abhishek Rudra
- *
  */
-
 @Entity
 @Table(name = "species_group")
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class SpeciesGroup implements Serializable {
 
-	/**
-	 * 
-	 */
+	/** */
 	private static final long serialVersionUID = 3399694424864046660L;
+
 	private Long id;
 	private String name;
 	private Long parentGroupId;
@@ -67,5 +62,4 @@ public class SpeciesGroup implements Serializable {
 	public void setGroupOrder(Integer groupOrder) {
 		this.groupOrder = groupOrder;
 	}
-
 }
