@@ -1,23 +1,19 @@
-/**
- * 
- */
+/** */
 package com.strandls.taxonomy.pojo;
 
 import java.util.Objects;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 
 /**
  * @author Abhishek Rudra
- *
- * 
  */
 @Entity
 @Table(name = "species_permission_request")
@@ -29,9 +25,7 @@ public class SpeciesPermissionRequest {
 	private Long userId;
 	private String role;
 
-	/**
-	 * 
-	 */
+	/** */
 	public SpeciesPermissionRequest() {
 		super();
 	}
@@ -87,7 +81,7 @@ public class SpeciesPermissionRequest {
 	public void setRole(String role) {
 		this.role = role;
 	}
-	
+
 	@Override
 	public int hashCode() {
 		return Objects.hash(id, role, taxonConceptId, userId);
@@ -124,5 +118,4 @@ public class SpeciesPermissionRequest {
 			return false;
 		return true;
 	}
-
 }
