@@ -1,23 +1,17 @@
-/**
- * 
- */
+/** */
 package com.strandls.taxonomy.pojo;
 
 import java.util.List;
 
 /**
  * @author Abhishek Rudra
- *
- * 
  */
 public class TaxonomicNames {
 
 	private List<CommonName> commonNames;
 	private List<TaxonomyDefinition> synonyms;
 
-	/**
-	 * 
-	 */
+	/** */
 	public TaxonomicNames() {
 		super();
 	}
@@ -47,5 +41,4 @@ public class TaxonomicNames {
 	public void setSynonyms(List<TaxonomyDefinition> synonyms) {
 		this.synonyms = synonyms;
 	}
-
 }
