@@ -1,79 +1,118 @@
-/**
- * 
- */
 package com.strandls.taxonomy.pojo;
 
 import java.sql.Timestamp;
 
-import javax.persistence.Column;
-import javax.persistence.ColumnResult;
-import javax.persistence.ConstructorResult;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Index;
-import javax.persistence.SequenceGenerator;
-import javax.persistence.SqlResultSetMapping;
-import javax.persistence.Table;
-
-import org.hibernate.annotations.Type;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.strandls.taxonomy.pojo.response.TaxonomyNamelistItem;
 
-/**
- * @author Abhishek Rudra
- *
- */
+import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.persistence.Column;
+import jakarta.persistence.ColumnResult;
+import jakarta.persistence.ConstructorResult;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Index;
+import jakarta.persistence.SequenceGenerator;
+import jakarta.persistence.SqlResultSetMapping;
+import jakarta.persistence.Table;
 
+@Schema(description = "Full Taxonomy Definition entity: canonical info, authorship, system mapping, and status")
 @SqlResultSetMapping(name = "TaxonomyNameList", classes = {
 		@ConstructorResult(targetClass = TaxonomyNamelistItem.class, columns = {
 				@ColumnResult(name = "id", type = Long.class), @ColumnResult(name = "rank", type = String.class),
 				@ColumnResult(name = "name", type = String.class), @ColumnResult(name = "status", type = String.class),
-				@ColumnResult(name = "position", type = String.class), @ColumnResult(name = "rankvalue", type = Double.class)
-				}) 
-		})
-
+				@ColumnResult(name = "position", type = String.class),
+				@ColumnResult(name = "rankvalue", type = Double.class) }) })
 @Entity
 @Table(name = "taxonomy_definition", indexes = {
 		@Index(name = "idx_canonical_form", columnList = "canonical_form, rank, is_deleted") })
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class TaxonomyDefinition {
 
+	@Schema(description = "Taxonomy unique ID", example = "12345")
 	private Long id;
+
+	@Schema(description = "Full binomial form, if applicable", example = "Homo sapiens")
 	private String binomialForm;
+
+	@Schema(description = "Canonical (normalized, stripped) form", required = true, example = "Homo sapiens")
 	private String canonicalForm;
+
+	@Schema(description = "Italicised form, for web display", required = true, example = "<i>Homo sapiens</i>")
 	private String italicisedForm;
+
+	@Schema(description = "External links ID for cross-ref", example = "56789")
 	private Long externalLinksId;
+
+	@Schema(description = "Original name as entered", required = true, example = "Homo sapiens Linnaeus, 1758")
 	private String name;
+
+	@Schema(description = "Normalized form for searching", example = "homo sapiens")
 	private String normalizedForm;
+
+	@Schema(description = "Rank e.g. SPECIES, GENUS", example = "SPECIES")
 	private String rank;
+
+	@Schema(description = "Timestamp of upload or update", example = "2024-07-29T10:15:00.0Z")
 	private Timestamp uploadTime;
+
+	@Schema(description = "Uploader User ID", example = "99")
 	private Long uploaderId;
+
+	@Schema(description = "Nomenclatural status", required = true, example = "ACCEPTED")
 	private String status;
+
+	@Schema(description = "Position (e.g. RAW, WORKING)", required = true, example = "RAW")
 	private String position;
+
+	@Schema(description = "Authorship and year", example = "Linnaeus, 1758")
 	private String authorYear;
+
+	@Schema(description = "External matched database name", example = "GBIF")
 	private String matchDatabaseName;
+
+	@Schema(description = "External match ID", example = "247")
 	private String matchId;
+
+	@Schema(description = "IBP Source", example = "IBP core")
 	private String ibpSource;
+
+	@Schema(description = "Via datasource (provenance)", example = "import")
 	private String viaDatasource;
+
+	@Schema(description = "Is flagged in curation?", example = "false")
 	private Boolean isFlagged;
+
+	@Schema(description = "Relationship string", example = "synonymOf")
 	private String relationship;
+
+	@Schema(description = "Special taxonomic class string", required = true, example = "MAMMALIA")
 	private String classs;
+
+	@Schema(description = "Reason for flagging (if any)", example = "Unresolved synonymy")
 	private String flaggingReason;
+
+	@Schema(description = "Is this name deleted?", example = "false")
 	private Boolean isDeleted;
+
+	@Schema(description = "Reason for being in dirty list", example = "Unparsed author")
 	private String dirtyListReason;
+
+	@Schema(description = "Description of curation activity", example = "Imported from legacy system")
 	private String activityDescription;
+
+	@Schema(description = "Default hierarchy (JSON or path)", example = "[1,2,3]")
 	private String defaultHierarchy;
+
+	@Schema(description = "Source system or provider identifier", example = "GBIF:1234")
 	private String nameSourceId;
 
 	public TaxonomyDefinition() {
-		super();
 	}
 
 	@Id
-	// @GeneratedValue
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "taxonomy_definition_id_generator")
 	@SequenceGenerator(name = "taxonomy_definition_id_generator", sequenceName = "taxonomy_definition_id_seq", allocationSize = 1)
 	@Column(name = "id")
@@ -194,7 +233,6 @@ public class TaxonomyDefinition {
 	}
 
 	@Column(name = "match_database_name")
-	@Type(type = "text")
 	public String getMatchDatabaseName() {
 		return matchDatabaseName;
 	}
@@ -222,7 +260,6 @@ public class TaxonomyDefinition {
 	}
 
 	@Column(name = "via_datasource")
-	@Type(type = "text")
 	public String getViaDatasource() {
 		return viaDatasource;
 	}
@@ -295,7 +332,6 @@ public class TaxonomyDefinition {
 	}
 
 	@Column(name = "default_hierarchy")
-	@Type(type = "text")
 	public String getDefaultHierarchy() {
 		return defaultHierarchy;
 	}
