@@ -1,48 +1,52 @@
-/**
- * 
- */
 package com.strandls.taxonomy.pojo;
 
 import java.io.Serializable;
 import java.sql.Timestamp;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.SequenceGenerator;
-import javax.persistence.Table;
-
-import org.hibernate.annotations.Type;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
-/**
- * @author Abhishek Rudra
- *
- */
+import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.SequenceGenerator;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "taxonomy_registry")
 @JsonIgnoreProperties(ignoreUnknown = true)
+@Schema(description = "Registry entry mapping a taxonomy node to a classification and path using PostgreSQL ltree")
 public class TaxonomyRegistry implements Serializable, Cloneable {
 
-	/**
-	 * 
-	 */
 	private static final long serialVersionUID = -1891934272853024930L;
-	
+
+	@Schema(description = "Registry ID")
 	private Long id;
+
+	@Schema(description = "Classification (tree) ID this registry belongs to")
 	private Long classificationId;
+
+	@Schema(description = "PostgreSQL ltree path string for fast ancestor/descendant queries")
 	private String path;
+
+	@Schema(description = "ID of the referenced taxonomy definition")
 	private Long taxonomyDefinationId;
+
+	@Schema(description = "Rank (e.g. species, genus)", required = true)
 	private String rank;
+
+	@Schema(description = "Timestamp of record upload/creation")
 	private Timestamp uploadTime;
+
+	@Schema(description = "User ID who uploaded/created the record")
 	private Long uploaderId;
 
+	public TaxonomyRegistry() {
+	}
+
 	@Id
-	//@GeneratedValue
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "taxonomy_registry_id_generator")
 	@SequenceGenerator(name = "taxonomy_registry_id_generator", sequenceName = "taxonomy_registry_id_seq", allocationSize = 1)
 	@Column(name = "id")
@@ -64,7 +68,6 @@ public class TaxonomyRegistry implements Serializable, Cloneable {
 	}
 
 	@Column(name = "path", columnDefinition = "ltree")
-	@Type(type = "com.strandls.taxonomy.pojo.enumtype.LTreeType")
 	public String getPath() {
 		return path;
 	}
@@ -86,7 +89,7 @@ public class TaxonomyRegistry implements Serializable, Cloneable {
 	public String getRank() {
 		return rank;
 	}
-	
+
 	public void setRank(String rank) {
 		this.rank = rank;
 	}
@@ -108,13 +111,9 @@ public class TaxonomyRegistry implements Serializable, Cloneable {
 	public void setUploaderId(Long uploaderId) {
 		this.uploaderId = uploaderId;
 	}
-	
+
 	@Override
 	public TaxonomyRegistry clone() throws CloneNotSupportedException {
-		Object object = super.clone();
-		if(object instanceof TaxonomyRegistry) {
-			return (TaxonomyRegistry) super.clone();
-		}
-		return this;
+		return (TaxonomyRegistry) super.clone();
 	}
 }
