@@ -6,10 +6,7 @@ import java.util.Map;
 import com.strandls.taxonomy.pojo.enumtype.TaxonomyStatus;
 
 /**
- * 
  * @author vilay
- *
- * 
  * @param taxonId        - Taxonomy id which need to be updated
  * @param taxonomyStatus - taxonomy status {@code TaxonomyStatus} to be updated
  * @param hierarchy      - This will be null for the accepted name to synonym
@@ -18,7 +15,6 @@ import com.strandls.taxonomy.pojo.enumtype.TaxonomyStatus;
  *                       change.
  * @return
  */
-
 public class TaxonomyStatusUpdate {
 
 	private Long taxonId;
@@ -30,7 +26,8 @@ public class TaxonomyStatusUpdate {
 		super();
 	}
 
-	public TaxonomyStatusUpdate(Long taxonId, TaxonomyStatus status, List<Long> newTaxonId, Map<String, String> hierarchy) {
+	public TaxonomyStatusUpdate(Long taxonId, TaxonomyStatus status, List<Long> newTaxonId,
+			Map<String, String> hierarchy) {
 		super();
 		this.taxonId = taxonId;
 		this.status = status;
@@ -69,5 +66,4 @@ public class TaxonomyStatusUpdate {
 	public void setHierarchy(Map<String, String> hierarchy) {
 		this.hierarchy = hierarchy;
 	}
-
 }
