@@ -18,13 +18,12 @@ public interface TaxonomyRegistryService {
 
 	public List<TaxonRelation> list(Long parent, String taxonIds, boolean expandTaxon, Long classificationId);
 
-	
-	/**
-	 * Code below from here on is only for the migration purpose
-	 */
+	/** Code below from here on is only for the migration purpose */
 	public Map<String, Object> migrateCleanName() throws CloneNotSupportedException;
 
-	public Map<String, Object> snapWorkingNames() throws CloneNotSupportedException, UnRecongnizedRankException, ApiException, TaxonCreationException;
+	public Map<String, Object> snapWorkingNames()
+			throws CloneNotSupportedException, UnRecongnizedRankException, ApiException, TaxonCreationException;
 
-	public Map<String, Object> snapRawNames() throws CloneNotSupportedException, UnRecongnizedRankException, ApiException, TaxonCreationException;
+	public Map<String, Object> snapRawNames()
+			throws CloneNotSupportedException, UnRecongnizedRankException, ApiException, TaxonCreationException;
 }
