@@ -3,24 +3,24 @@ package com.strandls.taxonomy.service.impl;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.inject.Inject;
-import javax.servlet.http.HttpServletRequest;
-
 import com.strandls.taxonomy.dao.RankDao;
 import com.strandls.taxonomy.pojo.Rank;
 import com.strandls.taxonomy.service.RankSerivce;
 import com.strandls.taxonomy.util.AbstractService;
 
+import jakarta.inject.Inject;
+import jakarta.servlet.http.HttpServletRequest;
+
 public class RankServiceImpl extends AbstractService<Rank> implements RankSerivce {
-	
+
 	@Inject
 	private RankDao rankDao;
-	
+
 	@Inject
 	public RankServiceImpl(RankDao dao) {
 		super(dao);
 	}
-	
+
 	@Override
 	public Rank fetchById(Long id) {
 		return rankDao.findById(id);
@@ -30,28 +30,28 @@ public class RankServiceImpl extends AbstractService<Rank> implements RankSerivc
 	public List<Rank> getAllRank(HttpServletRequest request) {
 		return rankDao.getAllRank();
 	}
-	
+
 	@Override
 	public List<String> getAllRankNames() {
 		List<Rank> ranks = rankDao.getAllRank();
 		List<String> rankNames = new ArrayList<>();
-		for(Rank rank : ranks) {
+		for (Rank rank : ranks) {
 			rankNames.add(rank.getName());
 		}
 		return rankNames;
 	}
-	
+
 	@Override
 	public List<String> getAllRequiredRanks() {
 		List<Rank> ranks = rankDao.getAllRank();
 		List<String> rankNames = new ArrayList<>();
-		for(Rank rank : ranks) {
-			if(rank.getIsRequired().booleanValue())
+		for (Rank rank : ranks) {
+			if (rank.getIsRequired().booleanValue())
 				rankNames.add(rank.getName());
 		}
 		return rankNames;
 	}
-	
+
 	@Override
 	public Rank addRequiredRank(HttpServletRequest request, String rankName, Double rankValue) {
 		Rank rank = new Rank();
@@ -65,16 +65,16 @@ public class RankServiceImpl extends AbstractService<Rank> implements RankSerivc
 	@Override
 	public Rank addIntermediateRank(HttpServletRequest request, String rankName, String highRankName,
 			String lowRankName) {
-		
+
 		Rank highRank = rankDao.findRankByName(highRankName);
 		Rank lowRank = rankDao.findRankByName(lowRankName);
-		
-		if(highRank == null || lowRank == null) 
+
+		if (highRank == null || lowRank == null)
 			return null;
-		
+
 		double rankValue = highRank.getRankValue() + lowRank.getRankValue();
 		rankValue /= 2.0;
-		
+
 		Rank rank = new Rank();
 		rank.setIsRequired(false);
 		rank.setName(rankName);
