@@ -7,8 +7,6 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 
-import javax.inject.Inject;
-
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.query.Query;
@@ -23,10 +21,12 @@ import com.strandls.taxonomy.TaxonomyConfig;
 import com.strandls.taxonomy.dao.TaxonomyDefinitionDao;
 import com.strandls.taxonomy.pojo.TaxonomyESDocument;
 
+import jakarta.inject.Inject;
+
 public class TaxonomyESOperation {
 
 	private final Logger logger = LoggerFactory.getLogger(TaxonomyESOperation.class);
-	
+
 	@Inject
 	private SessionFactory sessionFactory;
 
@@ -35,7 +35,7 @@ public class TaxonomyESOperation {
 
 	@Inject
 	private EsServicesApi esServicesApi;
-	
+
 	@Inject
 	private TaxonomyDefinitionDao taxonomyDefinitionDao;
 
@@ -43,16 +43,15 @@ public class TaxonomyESOperation {
 	private static final String ES_TAXONOMY_TYPE = "es.taxonomy.type";
 
 	private static final String SQL_FILE = "extendedTaxonDefinition.sql";
-	
+
 	private static final int FIXED_THREAD_SIZE = 10;
 	private static final int BATCH_SIZE = 1000;
 
 	private List<MapQueryResponse> esResult = new ArrayList<>();
-	
+
 	private static String qryString;
 
 	/**
-	 * 
 	 * @param taxonIds       - List of taxon Ids
 	 * @param createOrUpdate - true - If the taxonIds need to be created in the
 	 *                       elastic. false - If the taxonIds need to be updated in
@@ -68,24 +67,24 @@ public class TaxonomyESOperation {
 		}
 
 		int size = taxonIds.size();
-		
-		int numBatches = size/BATCH_SIZE;
-		
-		for(int i=0;i<numBatches;i++) {
-			List<Long> batch = taxonIds.subList(i*BATCH_SIZE, (i+1)*BATCH_SIZE);
+
+		int numBatches = size / BATCH_SIZE;
+
+		for (int i = 0; i < numBatches; i++) {
+			List<Long> batch = taxonIds.subList(i * BATCH_SIZE, (i + 1) * BATCH_SIZE);
 			ESThread esThread = new ESThread(this, batch);
 			executor.execute(esThread);
 		}
-		
+
 		// Execute for the last batch of taxonIds
-		List<Long> batch = taxonIds.subList(numBatches*BATCH_SIZE, size);
-		if(!batch.isEmpty()) {
+		List<Long> batch = taxonIds.subList(numBatches * BATCH_SIZE, size);
+		if (!batch.isEmpty()) {
 			ESThread esThread = new ESThread(this, batch);
 			executor.execute(esThread);
 		}
-		
+
 		executor.shutdown();
-		
+
 		try {
 			executor.awaitTermination(Long.MAX_VALUE, TimeUnit.NANOSECONDS);
 		} catch (InterruptedException e) {
@@ -94,36 +93,38 @@ public class TaxonomyESOperation {
 		}
 		return esResult;
 	}
-	
-	/** 
+
+	/**
 	 * This method is used by thread
+	 *
 	 * @param mapQueryResponses
 	 */
 	protected synchronized void updateResult(List<MapQueryResponse> mapQueryResponses) {
 		esResult.addAll(mapQueryResponses);
 	}
-	
+
 	/**
 	 * This method is used by thread to execute the elastic update for taxonomy
+	 *
 	 * @param taxonIds
 	 * @return
 	 */
 	protected List<MapQueryResponse> pushInBatches(List<Long> taxonIds) {
-		
+
 		Session session = sessionFactory.openSession();
 		session.beginTransaction();
-		
+
 		try {
 			Query<TaxonomyESDocument> query = session.createNativeQuery(qryString, TaxonomyESDocument.class);
 			query.setParameterList("taxonIds", taxonIds);
 			List<TaxonomyESDocument> taxonomyESDocuments = query.getResultList();
-			
+
 			String index = TaxonomyConfig.getString(ES_TAXONOMY_INDEX);
 			String type = TaxonomyConfig.getString(ES_TAXONOMY_TYPE);
-			
+
 			String taxonomyJsonData = objectMapper.writeValueAsString(taxonomyESDocuments);
 			return esServicesApi.bulkUpload(index, type, taxonomyJsonData);
-			
+
 		} catch (IOException e1) {
 			logger.error(e1.getMessage());
 		} catch (ApiException e) {
@@ -131,41 +132,39 @@ public class TaxonomyESOperation {
 		} finally {
 			session.close();
 		}
-		
+
 		return new ArrayList<>();
 	}
 
 	public List<MapQueryResponse> reIndexElastic() {
 		Long rowCount = taxonomyDefinitionDao.getRowCount();
 		int limit = BATCH_SIZE * 50;
-		
-		int batches = rowCount.intValue()/limit;
-		for(int i = 0 ;i <= batches; i++) {
+
+		int batches = rowCount.intValue() / limit;
+		for (int i = 0; i <= batches; i++) {
 			int offset = limit * i;
-			if(rowCount.intValue() < (offset + limit)) {
+			if (rowCount.intValue() < (offset + limit)) {
 				limit = rowCount.intValue() - offset;
 			}
 			List<Long> taxonIds = taxonomyDefinitionDao.getAllIds(limit, offset);
-			if(!taxonIds.isEmpty()) {
+			if (!taxonIds.isEmpty()) {
 				pushToElastic(taxonIds);
 			}
 		}
 		return esResult;
 	}
-
 }
 
+class ESThread extends TaxonomyESOperation implements Runnable {
 
-class ESThread extends TaxonomyESOperation implements Runnable{
-	
 	private List<Long> taxonIds;
 	private TaxonomyESOperation taxonomyESOperation;
-	
+
 	public ESThread(TaxonomyESOperation taxonomyESOperation, List<Long> taxonIds) {
 		this.taxonomyESOperation = taxonomyESOperation;
 		this.taxonIds = taxonIds;
 	}
-	
+
 	@Override
 	public void run() {
 		List<MapQueryResponse> mapQueryResponses = taxonomyESOperation.pushInBatches(taxonIds);
