@@ -1,6 +1,4 @@
-/**
- * 
- */
+/** */
 package com.strandls.taxonomy.util;
 
 import java.util.ArrayList;
@@ -8,8 +6,6 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
-import javax.inject.Inject;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -24,10 +20,10 @@ import com.strandls.mail_utility.util.JsonUtil;
 import com.strandls.taxonomy.RabbitMqConnection;
 import com.strandls.user.pojo.User;
 
+import jakarta.inject.Inject;
+
 /**
  * @author Abhishek Rudra
- *
- * 
  */
 public class MailUtils {
 
@@ -69,7 +65,6 @@ public class MailUtils {
 		} catch (Exception e) {
 			logger.error(e.getMessage());
 		}
-
 	}
 
 	public void sendPermissionGrant(User requestee, String taxonName, String role, Long taxonId) {
@@ -97,9 +92,6 @@ public class MailUtils {
 			} catch (Exception e) {
 				logger.error(e.getMessage());
 			}
-
 		}
-
 	}
-
 }
